# internal/platform/httpserver/openapi.yaml
openapi: 3.0.3
info:
  title: Backend CRM Piposmart
  version: 0.15.4-sprint-14f-identity
  description: API CRM Piposmart dengan Partner Type, Partner (Encrypted & Masked Bank Account), PIC Assignment, Call Interactions, Lead Referrals, Subscription Order, Reconciliation Sprint 11, Partner Commission Sprint 12, penambahan route trash/unscoped owner-outlet dan rekap status langganan outlet Sprint 14b/14c, histori import persisten dengan viewer/download file asli Sprint 14 history, route /all dan /all-deleted lintas modul untuk kebutuhan frontend mengambil full dataset tanpa pagination, serta penambahan manajemen akun Admin dan Supervisor beserta reset password lintas role pada Sprint 14f.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/leads:
    get:
      summary: List customer lead sesuai ownership actor
      operationId: listLeads
      description: Admin melihat semua lead, Supervisor melihat lead miliknya dan Sales di bawahnya, Sales hanya melihat lead miliknya.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
        - name: ownership
          in: query
          schema:
            type: string
            enum: [ADMIN, SUPERVISOR, SALES]
        - name: stage
          in: query
          schema:
            type: string
            example: POSSIBLE
        - name: status
          in: query
          schema:
            type: string
            example: OPEN
        - name: score
          in: query
          schema:
            type: integer
            enum: [0, 1, 2, 3]
        - name: supervisor_id
          in: query
          schema:
            type: integer
            format: int64
        - name: sales_id
          in: query
          schema:
            type: integer
            format: int64
        - name: follow_up_from
          in: query
          schema:
            type: string
            format: date
        - name: follow_up_to
          in: query
          schema:
            type: string
            format: date
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/LimitQuery"
      responses:
        "200":
          description: List lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
    post:
      summary: Membuat lead dari owner yang sudah ada
      operationId: createLead
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadRequest"
      responses:
        "201":
          description: Lead dibuat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /api/v1/leads/bulk/assign-supervisor:
    post:
      summary: Bulk assign lead dari Admin ke Supervisor
      operationId: bulkAssignSupervisor
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkAssignSupervisorRequest"
      responses:
        "200":
          description: Lead berhasil dibagikan ke Supervisor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /api/v1/leads/bulk/assign-sales:
    post:
      summary: Bulk assign lead dari Supervisor ke Sales
      operationId: bulkAssignSales
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkAssignSalesRequest"
      responses:
        "200":
          description: Lead berhasil dibagikan ke Sales
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /api/v1/leads/bulk/release:
    post:
      summary: Bulk release lead kembali ke Admin
      operationId: bulkReleaseLead
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkReleaseLeadRequest"
      responses:
        "200":
          description: Lead berhasil dikembalikan ke Admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /api/v1/leads/{lead_id}:
    get:
      summary: Detail customer lead
      operationId: getLead
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/LeadID"
      responses:
        "200":
          description: Detail lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /api/v1/leads/{lead_id}/assignment-history:
    get:
      summary: Riwayat perpindahan ownership lead
      operationId: getLeadAssignmentHistory
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/LeadID"
      responses:
        "200":
          description: Riwayat assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /api/v1/leads/{lead_id}/assign-supervisor:
    post:
      summary: Assign lead dari Admin ke Supervisor
      operationId: assignSupervisor
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/LeadID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignSupervisorRequest"
      responses:
        "200":
          description: Lead berhasil dibagikan ke Supervisor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /api/v1/leads/{lead_id}/assign-sales:
    post:
      summary: Assign atau redistribusi lead dari Supervisor ke Sales
      operationId: assignSales
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/LeadID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignSalesRequest"
      responses:
        "200":
          description: Lead berhasil dibagikan ke Sales
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /api/v1/leads/{lead_id}/release:
    post:
      summary: Release lead kembali ke Admin
      operationId: releaseLead
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/LeadID"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReleaseLeadRequest"
      responses:
        "200":
          description: Lead berhasil dikembalikan ke Admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /api/v1/leads/{lead_id}/mark-invalid:
    post:
      summary: Sales menandai lead invalid dari remark 0
      operationId: markLeadInvalid
      description: Lead berpindah kembali ke Supervisor, score menjadi 0, status menjadi INVALID.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/LeadID"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkInvalidRequest"
      responses:
        "200":
          description: Lead berhasil ditandai invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
...
  /api/v1/leads/all:
    get:
      summary: List semua lead tanpa pagination
      operationId: listLeadsAll
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
        - name: ownership
          in: query
          schema:
            type: string
        - name: stage
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: score
          in: query
          schema:
            type: integer
        - name: supervisor_id
          in: query
          schema:
            type: integer
            format: int64
        - name: sales_id
          in: query
          schema:
            type: integer
            format: int64
        - name: follow_up_from
          in: query
          schema:
            type: string
            format: date
        - name: follow_up_to
          in: query
          schema:
            type: string
            format: date
        - name: sort
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List semua lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /api/v1/leads/all-deleted:
    get:
      summary: Alias full lead untuk kompatibilitas frontend
      operationId: listLeadsAllDeleted
      description: Modul lead saat ini tidak memakai soft delete. Route ini disediakan sebagai alias kompatibilitas frontend.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
        - name: ownership
          in: query
          schema:
            type: string
        - name: stage
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: score
          in: query
          schema:
            type: integer
        - name: supervisor_id
          in: query
          schema:
            type: integer
            format: int64
        - name: sales_id
          in: query
          schema:
            type: integer
            format: int64
        - name: follow_up_from
          in: query
          schema:
            type: string
            format: date
        - name: follow_up_to
          in: query
          schema:
            type: string
            format: date
        - name: sort
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List semua lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    LeadID:
      name: lead_id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    SearchQuery:
      name: q
      in: query
      required: false
      schema:
        type: string
      description: Search global untuk kode, nama, telepon, brand, kota, dan provinsi.
...
    PageQuery:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 1
        minimum: 1
    LimitQuery:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        default: 10
        minimum: 1
        maximum: 100
  schemas:
...
    CreateLeadRequest:
      type: object
      required: [owner_id]
      properties:
        owner_id:
          type: integer
          format: int64
        outlet_id:
          type: integer
          format: int64
          nullable: true
        source_type:
          type: string
          example: MANUAL
        source_reference:
          type: string
    AssignSupervisorRequest:
      type: object
      required: [supervisor_id]
      properties:
        supervisor_id:
          type: integer
          format: int64
        reason:
          type: string
          example: Distribusi dari Admin ke Supervisor
    BulkAssignSupervisorRequest:
      type: object
      required: [lead_ids, supervisor_id]
      properties:
        lead_ids:
          type: array
          minItems: 1
          items:
            type: integer
            format: int64
        supervisor_id:
          type: integer
          format: int64
        reason:
          type: string
    AssignSalesRequest:
      type: object
      required: [sales_id]
      properties:
        sales_id:
          type: integer
          format: int64
        reason:
          type: string
          example: Distribusi dari Supervisor ke Sales
    BulkAssignSalesRequest:
      type: object
      required: [lead_ids, sales_id]
      properties:
        lead_ids:
          type: array
          minItems: 1
          items:
            type: integer
            format: int64
        sales_id:
          type: integer
          format: int64
        reason:
          type: string
    ReleaseLeadRequest:
      type: object
      properties:
        admin_id:
          type: integer
          format: int64
          nullable: true
        reason:
          type: string
          example: Dikembalikan ke Admin
    BulkReleaseLeadRequest:
      type: object
      required: [lead_ids]
      properties:
        lead_ids:
          type: array
          minItems: 1
          items:
            type: integer
            format: int64
        admin_id:
          type: integer
          format: int64
          nullable: true
        reason:
          type: string
    MarkInvalidRequest:
      type: object
      properties:
        reason:
          type: string
          example: Customer tidak potensial dari call terakhir
...
    SuccessEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          additionalProperties: true
        meta:
          type: object
          required: [request_id]
          properties:
            request_id:
              type: string
...